# web/openapi.json
{
  "components": {
    "schemas": {
      "AgentSessionSummarySchema": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "message_count": {
            "default": 0,
            "title": "Message Count",
            "type": "integer"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          },
          "session_type": {
            "$ref": "#/components/schemas/SessionType",
            "default": "chat"
          },
          "title": {
            "default": "",
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "session_id",
          "created_at",
          "updated_at"
        ],
        "title": "AgentSessionSummarySchema",
        "type": "object"
      },
...
      "CommonResponse": {
        "properties": {
          "code": {
            "default": 200,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Data"
          },
          "message": {
            "default": "success",
            "title": "Message",
            "type": "string"
          }
        },
        "title": "CommonResponse",
        "type": "object"
      },
      "CommonResponse_Any_": {
        "properties": {
          "code": {
            "default": 200,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "title": "Data"
          },
          "message": {
            "default": "success",
            "title": "Message",
            "type": "string"
          }
        },
        "title": "CommonResponse[Any]",
        "type": "object"
      },
      "CommonResponse_CreateAgentSessionResponse_": {
        "properties": {
          "code": {
            "default": 200,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CreateAgentSessionResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "default": "success",
            "title": "Message",
            "type": "string"
          }
        },
        "title": "CommonResponse[CreateAgentSessionResponse]",
        "type": "object"
      },
      "CommonResponse_DeleteSandboxImageResponse_": {
        "properties": {
          "code": {
            "default": 200,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DeleteSandboxImageResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "default": "success",
            "title": "Message",
            "type": "string"
          }
        },
        "title": "CommonResponse[DeleteSandboxImageResponse]",
        "type": "object"
      },
...
      "CommonResponse_ListAgentEventsResponse_": {
        "properties": {
          "code": {
            "default": 200,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ListAgentEventsResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "default": "success",
            "title": "Message",
            "type": "string"
          }
        },
        "title": "CommonResponse[ListAgentEventsResponse]",
        "type": "object"
      },
      "CommonResponse_ListAgentSessionsResponse_": {
        "properties": {
          "code": {
            "default": 200,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ListAgentSessionsResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "default": "success",
            "title": "Message",
            "type": "string"
          }
        },
        "title": "CommonResponse[ListAgentSessionsResponse]",
        "type": "object"
      },
      "CommonResponse_QuerySandboxImagesResponse_": {
        "properties": {
          "code": {
            "default": 200,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/QuerySandboxImagesResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "default": "success",
            "title": "Message",
            "type": "string"
          }
        },
        "title": "CommonResponse[QuerySandboxImagesResponse]",
        "type": "object"
      },
...
      "CommonResponse_SandboxImageSchema_": {
        "properties": {
          "code": {
            "default": 200,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SandboxImageSchema"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "default": "success",
            "title": "Message",
            "type": "string"
          }
        },
        "title": "CommonResponse[SandboxImageSchema]",
        "type": "object"
      },
...
      "CreateAgentSessionResponse": {
        "properties": {
          "session_id": {
            "title": "Session Id",
            "type": "string"
          }
        },
        "required": [
          "session_id"
        ],
        "title": "CreateAgentSessionResponse",
        "type": "object"
      },
      "CreateSandboxImageRequest": {
        "properties": {
          "image_name": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Image Name",
            "type": "string"
          }
        },
        "required": [
          "image_name"
        ],
        "title": "CreateSandboxImageRequest",
        "type": "object"
      },
...
      "DeleteSandboxImageResponse": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "integer"
          }
        },
        "required": [
          "id"
        ],
        "title": "DeleteSandboxImageResponse",
        "type": "object"
      },
...
      "ErrorEvent": {
        "properties": {
          "agent_name": {
            "default": "",
            "title": "Agent Name",
            "type": "string"
          },
          "code": {
            "default": "",
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "const": "error",
            "default": "error",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "message"
        ],
        "title": "ErrorEvent",
        "type": "object"
      },
      "HandoffEvent": {
        "properties": {
          "source_agent": {
            "title": "Source Agent",
            "type": "string"
          },
          "target_agent": {
            "title": "Target Agent",
            "type": "string"
          },
          "type": {
            "const": "handoff",
            "default": "handoff",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "source_agent",
          "target_agent"
        ],
        "title": "HandoffEvent",
        "type": "object"
      },
      "ListAgentEventsResponse": {
        "properties": {
          "items": {
            "items": {
              "discriminator": {
                "mapping": {
                  "error": "#/components/schemas/ErrorEvent",
                  "handoff": "#/components/schemas/HandoffEvent",
                  "text_complete": "#/components/schemas/TextCompleteEvent",
                  "text_delta": "#/components/schemas/TextDeltaEvent",
                  "thinking_complete": "#/components/schemas/ThinkingCompleteEvent",
                  "thinking_delta": "#/components/schemas/ThinkingDeltaEvent",
                  "tool_call": "#/components/schemas/ToolCallEvent",
                  "tool_result": "#/components/schemas/ToolResultEvent",
                  "user_message": "#/components/schemas/UserMessageEvent"
                },
                "propertyName": "type"
              },
              "oneOf": [
                {
                  "$ref": "#/components/schemas/UserMessageEvent"
                },
                {
                  "$ref": "#/components/schemas/TextDeltaEvent"
                },
                {
                  "$ref": "#/components/schemas/TextCompleteEvent"
                },
                {
                  "$ref": "#/components/schemas/ThinkingDeltaEvent"
                },
                {
                  "$ref": "#/components/schemas/ThinkingCompleteEvent"
                },
                {
                  "$ref": "#/components/schemas/ToolCallEvent"
                },
                {
                  "$ref": "#/components/schemas/ToolResultEvent"
                },
                {
                  "$ref": "#/components/schemas/HandoffEvent"
                },
                {
                  "$ref": "#/components/schemas/ErrorEvent"
                }
              ]
            },
            "title": "Items",
            "type": "array"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          }
        },
        "required": [
          "session_id",
          "items"
        ],
        "title": "ListAgentEventsResponse",
        "type": "object"
      },
      "ListAgentSessionsResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AgentSessionSummarySchema"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "ListAgentSessionsResponse",
        "type": "object"
      },
      "QuerySandboxImagesResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/SandboxImageSchema"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "size": {
            "title": "Size",
            "type": "integer"
          }
        },
        "required": [
          "page",
          "size",
          "items"
        ],
        "title": "QuerySandboxImagesResponse",
        "type": "object"
      },
...
      "SandboxImageSchema": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "image_hash": {
            "title": "Image Hash",
            "type": "string"
          },
          "image_name": {
            "title": "Image Name",
            "type": "string"
          },
          "image_size": {
            "title": "Image Size",
            "type": "integer"
          },
          "status": {
            "$ref": "#/components/schemas/SandboxImageStatus"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "image_name",
          "image_size",
          "image_hash",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "SandboxImageSchema",
        "type": "object"
      },
      "SandboxImageStatus": {
        "enum": [
          "pulling",
          "ready",
          "failed",
          "canceled"
        ],
        "title": "SandboxImageStatus",
        "type": "string"
      },
...
      "TextCompleteEvent": {
        "properties": {
          "agent_name": {
            "default": "",
            "title": "Agent Name",
            "type": "string"
          },
          "item_id": {
            "title": "Item Id",
            "type": "string"
          },
          "text": {
            "title": "Text",
            "type": "string"
          },
          "type": {
            "const": "text_complete",
            "default": "text_complete",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "item_id",
          "text"
        ],
        "title": "TextCompleteEvent",
        "type": "object"
      },
      "TextDeltaEvent": {
        "properties": {
          "agent_name": {
            "default": "",
            "title": "Agent Name",
            "type": "string"
          },
          "delta": {
            "title": "Delta",
            "type": "string"
          },
          "item_id": {
            "title": "Item Id",
            "type": "string"
          },
          "type": {
            "const": "text_delta",
            "default": "text_delta",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "item_id",
          "delta"
        ],
        "title": "TextDeltaEvent",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Z3r0 - AI + Security Attack Platform",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/agent-sessions": {
      "get": {
        "operationId": "list_agent_sessions_route_api_agent_sessions_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_ListAgentSessionsResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "List Agent Sessions Route",
        "tags": [
          "agent-sessions"
        ]
      },
      "post": {
        "operationId": "create_agent_session_handler_api_agent_sessions_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_CreateAgentSessionResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Create Agent Session Handler",
        "tags": [
          "agent-sessions"
        ]
      }
    },
    "/api/agent-sessions/{session_id}": {
      "delete": {
        "operationId": "delete_agent_session_handler_api_agent_sessions__session_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Agent session not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Delete Agent Session Handler",
        "tags": [
          "agent-sessions"
        ]
      }
    },
    "/api/agent-sessions/{session_id}/events": {
      "get": {
        "operationId": "list_agent_events_route_api_agent_sessions__session_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_ListAgentEventsResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "List Agent Events Route",
        "tags": [
          "agent-sessions"
        ]
      }
    },
    "/api/sandbox-images": {
      "get": {
        "operationId": "query_sandbox_images_route_api_sandbox_images_get",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "size",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 100,
              "minimum": 1,
              "title": "Size",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "keyword",
            "required": false,
            "schema": {
              "default": "",
              "title": "Keyword",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_QuerySandboxImagesResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Forbidden"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Query Sandbox Images Route",
        "tags": [
          "sandbox-images"
        ]
      },
      "post": {
        "operationId": "create_sandbox_image_handler_api_sandbox_images_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSandboxImageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_SandboxImageSchema_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Forbidden"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Create Sandbox Image Handler",
        "tags": [
          "sandbox-images"
        ]
      }
    },
    "/api/sandbox-images/{id}": {
      "delete": {
        "operationId": "delete_sandbox_image_handler_api_sandbox_images__id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_DeleteSandboxImageResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Sandbox image not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Delete Sandbox Image Handler",
        "tags": [
          "sandbox-images"
        ]
      }
    },
    "/api/sandbox-images/{id}/cancel": {
      "post": {
        "operationId": "cancel_sandbox_image_pull_handler_api_sandbox_images__id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_SandboxImageSchema_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Sandbox image not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Cancel Sandbox Image Pull Handler",
        "tags": [
          "sandbox-images"
        ]
      }
    },
    "/api/sandbox-images/{id}/retry": {
      "post": {
        "operationId": "retry_sandbox_image_handler_api_sandbox_images__id__retry_post",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_SandboxImageSchema_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Sandbox image not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Retry Sandbox Image Handler",
        "tags": [
          "sandbox-images"
        ]
      }
    },
...
  }
}